refactor: 중복되는 youtube type 추가/설정 및 삭제
