+++ apps/admin/src/types/youtube/response.ts
@@ -1,8 +1,3 @@
-import { IYoutube } from "type";
-
-export interface IGetYoutubeResponse {
-  youtubeList: IYoutube[];
-}
 export interface IPostYoutubeResponse {
   result: string;
 }
