+++ apps/client/src/types/youtube/response.ts
@@ -1,6 +1,3 @@
-export interface IGetYoutubeResponse {
-  youtubeLink: string[];
-}
 export interface IPostYoutubeResponse {
   result: string;
 }
