+++ packages/type/src/index.ts
@@ -1,5 +1,11 @@
 import { IBulletin, IBulletinForm, IBulletinImageForm } from "./bulletin";
-import { YoutubeType, IYoutube, IYoutubeForm } from "./youtube";
+import {
+  YoutubeType,
+  IYoutube,
+  IYoutubeForm,
+  IGetYoutubeResponse,
+  IGetYoutubeListProps,
+} from "./youtube";
 
 export type {
   IBulletin,
@@ -8,4 +14,6 @@ export type {
   YoutubeType,
   IYoutube,
   IYoutubeForm,
+  IGetYoutubeResponse,
+  IGetYoutubeListProps,
 };
